refactor: restcontroller에 대한 유효성 검사 수행
