+++ src/main/java/gift/controller/MemberController.java
@@ -3,13 +3,14 @@ package gift.controller;
 import gift.dto.MemberRequestDto;
 import gift.service.JwtAuthService;
 import gift.service.MemberService;
+import jakarta.validation.Valid;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
 @RestController
 @RequestMapping("/api/members")
 public class MemberController {
@@ -25,7 +26,7 @@ public class MemberController {
 
     //TODO: 회원가입 기능 -> 토큰을 반환
     @PostMapping("/register")
-    public ResponseEntity<String> register(@RequestBody MemberRequestDto memberRequestDto){
+    public ResponseEntity<String> register(@RequestBody @Valid MemberRequestDto memberRequestDto){
         if(memberService.register(memberRequestDto)){
             String token = jwtAuthService.createJwt(memberRequestDto);
             return new ResponseEntity<>(token, HttpStatus.CREATED);
@@ -36,7 +37,7 @@ public class MemberController {
     //TODO: 로그인 기능 -> 토큰을 반환
     @PostMapping("/login")
     public ResponseEntity<String> login(
-            @RequestBody  MemberRequestDto memberRequestDto
+            @RequestBody @Valid MemberRequestDto memberRequestDto
     ){
         //서버에 저장된 id-pw 쌍과 일치하는지 확인
         if(!memberService.checkMember(memberRequestDto)){
